move tests to regression/issue/12041; carry had_too_recent on FoundLatestFallback
The pre-existing tests in bun-update.test.ts spawn multiple subprocesses
and sit at the edge of the 5s default timeout under ASAN, which flakes
in constrained environments. Move the new tests to their own file and
drop the separate install step (bun update --latest on a fresh
package.json exercises the same resolve path).

Also carry had_too_recent on FoundLatestFallback so latest_is_filtered()
reports it, and assert the fallback warning in the no-dist-tags test.
