+++ src/install/PackageManager/PackageManagerEnqueue.rs
@@ -2390,7 +2390,7 @@ fn get_or_put_resolved_package(
 
             let find_result_opt: Option<Npm::FindResult> = match version_result {
                 Npm::FindVersionResult::Found(result) => Some(result),
-                Npm::FindVersionResult::FoundLatestFallback(result) => {
+                Npm::FindVersionResult::FoundLatestFallback { result, .. } => {
                     let package_name = this.lockfile.str(&name);
                     this.log_mut().add_warning_fmt(
                         None,
