+++ test/cli/install/bun-update.test.ts
@@ -527,186 +527,3 @@ it("should print UTF-8 arrows correctly with colors enabled", async () => {
   expect(out).not.toContain("â");
   expect(await exited2).toBe(0);
 });
-
-// https://github.com/oven-sh/bun/issues/12041
-// Registry mirrors/CDNs can briefly serve a manifest whose `dist-tags.latest`
-// points at a version that has not yet propagated into `versions`. `bun update
-// --latest` should fall back to the highest published version rather than
-// aborting the whole update.
-it("should fall back to the highest version when dist-tags.latest is not in versions", async () => {
-  const urls: string[] = [];
-  // dist-tags.latest = 0.0.7 but only 0.0.3 and 0.0.5 are in `versions`
-  setHandler(
-    dummyRegistry(urls, {
-      "0.0.3": { bin: { "baz-run": "index.js" } },
-      "0.0.5": { bin: { "baz-exec": "index.js" } },
-      latest: "0.0.7",
-    }),
-  );
-  await writeFile(
-    join(package_dir, "package.json"),
-    JSON.stringify({
-      name: "foo",
-      dependencies: { baz: "0.0.3" },
-    }),
-  );
-
-  const install = spawn({
-    cmd: [bunExe(), "install", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [, installErr, installExit] = await Promise.all([
-    new Response(install.stdout).text(),
-    new Response(install.stderr).text(),
-    install.exited,
-  ]);
-  expect(installErr).not.toContain("error:");
-  expect(installExit).toBe(0);
-
-  const { stdout, stderr, exited } = spawn({
-    cmd: [bunExe(), "update", "--latest", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [out, err, exitCode] = await Promise.all([new Response(stdout).text(), new Response(stderr).text(), exited]);
-
-  expect(err).not.toContain('with tag "latest" not found');
-  expect(err).not.toContain("failed to resolve");
-  expect(err).toContain("warn:");
-  expect(err).toContain("falling back to 0.0.5");
-  expect(out).toContain("baz 0.0.3 -> 0.0.5");
-  expect(exitCode).toBe(0);
-
-  expect(await file(join(package_dir, "node_modules", "baz", "package.json")).json()).toMatchObject({
-    name: "baz",
-    version: "0.0.5",
-  });
-});
-
-it("should fall back to the highest version when the manifest has no dist-tags", async () => {
-  const urls: string[] = [];
-  setHandler(async request => {
-    urls.push(request.url);
-    const url = request.url.replaceAll("%2f", "/");
-    if (url.endsWith(".tgz")) {
-      return new Response(file(join(import.meta.dir, "baz-0.0.3.tgz")));
-    }
-    const name = url.slice(url.indexOf("/", root_url.length) + 1);
-    return new Response(
-      JSON.stringify({
-        name,
-        versions: {
-          "0.0.3": {
-            name,
-            version: "0.0.3",
-            bin: { "baz-run": "index.js" },
-            dist: { tarball: `${root_url}/${name}-0.0.3.tgz` },
-          },
-        },
-        // no "dist-tags"
-      }),
-    );
-  });
-  await writeFile(
-    join(package_dir, "package.json"),
-    JSON.stringify({
-      name: "foo",
-      dependencies: { baz: "0.0.3" },
-    }),
-  );
-
-  const install = spawn({
-    cmd: [bunExe(), "install", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [, installErr, installExit] = await Promise.all([
-    new Response(install.stdout).text(),
-    new Response(install.stderr).text(),
-    install.exited,
-  ]);
-  expect(installErr).not.toContain("error:");
-  expect(installExit).toBe(0);
-
-  const { stdout, stderr, exited } = spawn({
-    cmd: [bunExe(), "update", "--latest", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [out, err, exitCode] = await Promise.all([new Response(stdout).text(), new Response(stderr).text(), exited]);
-
-  expect(err).not.toContain('with tag "latest" not found');
-  expect(err).not.toContain("failed to resolve");
-  expect(exitCode).toBe(0);
-
-  expect(await file(join(package_dir, "node_modules", "baz", "package.json")).json()).toMatchObject({
-    name: "baz",
-    version: "0.0.3",
-  });
-});
-
-it("should fall back to the highest prerelease when the manifest has no releases and no dist-tags", async () => {
-  setHandler(async request => {
-    const url = request.url.replaceAll("%2f", "/");
-    if (url.endsWith(".tgz")) {
-      return new Response(file(join(import.meta.dir, "baz-0.0.3.tgz")));
-    }
-    const name = url.slice(url.indexOf("/", root_url.length) + 1);
-    const pre = (v: string) => ({ name, version: v, dist: { tarball: `${root_url}/${name}-${v}.tgz` } });
-    return new Response(
-      JSON.stringify({
-        name,
-        versions: {
-          "1.0.0-beta.1": pre("1.0.0-beta.1"),
-          "1.0.0-beta.2": pre("1.0.0-beta.2"),
-        },
-        // no "dist-tags", no stable releases
-      }),
-    );
-  });
-  await writeFile(
-    join(package_dir, "package.json"),
-    JSON.stringify({ name: "foo", dependencies: { baz: "1.0.0-beta.1" } }),
-  );
-
-  const install = spawn({
-    cmd: [bunExe(), "install", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [, installErr, installExit] = await Promise.all([
-    new Response(install.stdout).text(),
-    new Response(install.stderr).text(),
-    install.exited,
-  ]);
-  expect(installErr).not.toContain("error:");
-  expect(installExit).toBe(0);
-
-  const { stdout, stderr, exited } = spawn({
-    cmd: [bunExe(), "update", "--latest", "--linker=hoisted"],
-    cwd: package_dir,
-    stdout: "pipe",
-    stderr: "pipe",
-    env,
-  });
-  const [out, err, exitCode] = await Promise.all([new Response(stdout).text(), new Response(stderr).text(), exited]);
-
-  expect(err).not.toContain('with tag "latest" not found');
-  expect(err).not.toContain("failed to resolve");
-  expect(err).toContain("falling back to 1.0.0-beta.2");
-  expect(out).toContain("baz 1.0.0-beta.1 -> 1.0.0-beta.2");
-  expect(exitCode).toBe(0);
-
-  expect((await file(join(package_dir, "package.json")).json()).dependencies.baz).toContain("1.0.0-beta.2");
-});
